Count UltraFeed thread repetition once per card exposure
